+++ lib/src/mixin.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'store.dart';
+import 'provider.dart';
+
+mixin IConnect {
+  dynamic model;
+  dynamic register([VoidCallback? dispose]) {
+    return Store().register(this, dispose);
+  }
+
+  dynamic unRegister([VoidCallback? dispose]) {
+    return Store().unRegister(this, dispose);
+  }
+
+  listen(context) {
+    return Provider.of(context, this);
+  }
+
+  dispatch([Function? action]) {
+    return Store.instance.dispatch(this, action);
+  }
+}